Separated out game from game base for testing browser
